Establish a bare minimum requirements file for python deps
These are the requirements required to make the service run.
Many of these are not actually required, but because of the
way some packages import (notably nova.api.openstack importing
nova.utils) there are lots of unused extras. Some of those
extras require binary packages (loaded by apk).

We install the dependent reqs, and then install nova with
--no-deps to avoid installing all the usual requirements.