Switch to using TCP for the uwsgi socket
This makes it easier to have the containers that contain placement,
anywhere. We bind to port 80 local to the container.

For a running config, with mod_proxy_uwsgi we need to enable
'proxy_balancer' and 'lbmethod_byrequests' and configure apache as
follows:

    <Proxy balancer://placement>
       BalancerMember uwsgi://127.0.0.1:8080/
       BalancerMember uwsgi://127.0.0.1:8081/
    </Proxy>

    ProxyPass "/placement" balancer://placement

Once the contain is built we can run two of them, exposing each's
port 80 to 8080 or 8081:

for i in 8080 8081 ; do \
    sudo docker run \
       -dt -p 127.0.0.1:$i:80 \
       -v /home/cdent/src/placedock/shared:/shared cdent/placementdock ; \
done

This is not yet using the improved nova.utils found in
https://review.openstack.org/#/c/533797/ to limit impmorts, and
unfortunately it won't make a huge amount of different: the registration
of nova.objects imports many things, including cursive, which gets
castellan and cryptography and much else.