Hack up a standalone placement service
This builds the database within the container using sqlite as the
engine. This does (intentionally) mean that the data dies when the
container dies.

Needing to do database migrations (to create the tables) means that we
have to import a lot of stuff that was not required before. This can
probably be tuned with time, but for now all we're after is something
quick and dirty.

Also, switch off keystone and use the NoAuth middleware so we can avoid
keystone. Requests that looks like this will be accepted as authed:

curl -v -H "x-auth-token: admin" http://ds1/placement/resource_providers