Hack sync.py to limit imports
It turns out that there are many steps in in nova's migration db_sync
that can be skipped it we just do things ourselves: create the tables,
sync the migrations.

However, one migration, for keypairs, decides to import
nova.objects.keypair to get access to a single string constant. This
simple action ends up importing the world, but it means all the objects
get registered and everything they import is imported.

Since keypairs are not used in placement we can ignore that migration
but simply editing the file won't work because the system is looking for
a numerically ordered list. So we overwrite the migration file with
an upgrade method that does a 'pass'.

With this in place we no longer need to import cursive or
os-service-types and don't need libxml related packages.

And we can finally run a placement service with it's own data.