Update to get things working with kubernetes
This is mostly for the sake of learning: the fact that the sqlite
database is in container means it's fairly useless: you can't scale
horizontally. Future work will explore different ways of dealing
what that.

The Dockerfile is update to reflect the latest placement extraction
changes and to put config in the container rather than using a
shared volume.

The placement-uwsgi.ini is update so it runs http directly.
There's no need for an nginx or apache setup, a k8s LoadBalancer
in deployment.yaml takes care of that.

The uwsgi ini is based on info at:
https://github.com/nginxinc/kubernetes-ingress/issues/143#issuecomment-347814243

Using minikube the steps are:

    eval $(minikube docker-env)
    docker build -t placedock:1.0 .
    kubectl apply -f deployment.yaml
    kubectl expose deployment placement-deployment --type=LoadBalancer
    minikube service placement-deployment --url