Make the database configurable by environment variable
And do the database sync at run time, not build time.

This makes it possible to build the plain container and
then at run time delare that you want to point it to mysql
somewhere. With that, we can have more than 1 replica and
establish autoscaling.

So far this has only been testing with a failing non-default,
to confirm that the substution does indeed happen.