Enable use of an external postgres database
By default it is still sqlite in the container, but if
deployment.yaml is edited (see the comments in it) it is possible
to run with replicas, all of them using the same database string.

The Dockerfile has been updated to include psycopg2, so connections
to postgresql are possible. If other drivers supported by
SQLAlchemy they should work too.