Enable autoscaling
This suite of changes enables autoscaling, by turning on the
metrics service and API needed by recent k8s. This gets rid
of the need for the heapster addon, which didn't want to
work with k8s v1.9.0 with excessive tweaks.

The metrics*.yaml files are borrowed from minikube master.
When something beyond minikube 0.25.0 is released, these
will no longer be needed: the metrics server will be
available as an addon.

The autoscaler.yaml is adapted from the one created when
using the kubectl autoscale command.

sync.py is adjusted to ignore an error when it tries to
sync a database that has already been managed.

This will get written up on https://anticdent.org/ but
the gist is that assuming you connect to an external database,
you can start up placement with just one container, and then
after slamming it will requests it will expand and contract
the number of containers as load on the individual containeers
demands. It's not yet clear how quickly the metrics are
gathered or scaling changes happen. Will investigate that
further.