Extend use of environment variable for controlling use
The amount of templating in nova.conf is increased to allow keystone
auth to be turned on and configured.

This brings us closer to being able to create a container that either
can be used:

* with internal or external database
* syncing or not syncing database
* use keystone or noauth2

Which makes it easier to work in devstack or k8s environments depending
on requirements.