Use oslo.config 6.7.0
6.7.0 supports getting configuration from the environment so
this change removes the infrastructure to manage and modify
the placement.conf file, replaced with environment variable
setting.

Since 'placement-manage db sync' is now available the sync.py
command is no longer needed.

The README.md is updated to reflect the changes that have
happened with placement since it was extracted from nova.