Use oslo.config 6.7.0
6.7.0 supports getting configuration from the environment so
this change removes the infrastructure to manage and modify
the placement.conf file, replaced with environment variable
setting.

This is working fine with a local db and the noauth2 strategy
but has not yet been testing with a remote db and keystone. The
changes are in place, but nothing to test against at the moment.